get_pages_lambda/requirements.txt
attrs==20.3.0
aws-cdk.assets==1.106.1
aws-cdk.aws-applicationautoscaling==1.106.1
aws-cdk.aws-autoscaling-common==1.106.1
aws-cdk.aws-cloudwatch==1.106.1
aws-cdk.aws-codeguruprofiler==1.106.1
aws-cdk.aws-ec2==1.106.1
aws-cdk.aws-ecr==1.106.1
aws-cdk.aws-ecr-assets==1.106.1
aws-cdk.aws-efs==1.106.1
aws-cdk.aws-events==1.106.1
aws-cdk.aws-iam==1.106.1
aws-cdk.aws-kms==1.106.1
aws-cdk.aws-lambda==1.106.1
aws-cdk.aws-lambda-python==1.106.1
aws-cdk.aws-logs==1.106.1
aws-cdk.aws-rds==1.106.1
aws-cdk.aws-s3==1.106.1
aws-cdk.aws-s3-assets==1.106.1
aws-cdk.aws-sam==1.106.1
aws-cdk.aws-secretsmanager==1.106.1
aws-cdk.aws-signer==1.106.1
aws-cdk.aws-sns==1.106.1
aws-cdk.aws-sqs==1.106.1
aws-cdk.aws-ssm==1.106.1
aws-cdk.cloud-assembly-schema==1.106.1
aws-cdk.core==1.106.1
aws-cdk.cx-api==1.106.1
aws-cdk.region-info==1.106.1
boto3==1.17.85
botocore==1.20.85
cattrs==1.5.0
constructs==3.3.75
jmespath==0.10.0
jsii==1.30.0
logger==1.4
mysql-connector==2.2.9
publication==0.0.3
python-dateutil==2.8.1
s3transfer==0.4.2
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.5
